Refactor type Product; add product's route
