+++ lib/types/product.ts
@@ -1,0 +1,12 @@
+export type Product = {
+  id: number;
+  title: string;
+  price: number | string;
+  description: string;
+  category: string;
+  image: string;
+  rating: {
+    rate: number;
+    count: number;
+  };
+};
